Refactor: int 사용하던 것을 Long으로 수정

+++ src/main/java/codesquad/codestagram/controller/ArticleController.java
@@ -6,7 +6,6 @@ import codesquad.codestagram.dto.ArticleForm;
 import codesquad.codestagram.service.ArticleService;
 import codesquad.codestagram.utility.TextUtility;
 import jakarta.servlet.http.HttpSession;
-import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -42,7 +41,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{index}")
-    public String viewArticle(@PathVariable int index, Model model, HttpSession session) {
+    public String viewArticle(@PathVariable Long index, Model model, HttpSession session) {
         Article article = articleService.findArticleById(index);
         model.addAttribute("article", article);
         model.addAttribute("parsedContent", TextUtility.escapeAndConvertNewlines(article.getContent()));
@@ -50,7 +49,7 @@ public class ArticleController {
     }
 
     @DeleteMapping("/articles/{index}")
-    public String delete(@PathVariable int index, HttpSession session) {
+    public String delete(@PathVariable Long index, HttpSession session) {
         User loginUser = (User) session.getAttribute("loginUser");
         Article article = articleService.findArticleIfOwner(loginUser, index);
         articleService.delete(article);
@@ -58,7 +57,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{index}/update-form")
-    public String updateForm(@PathVariable int index, Model model, HttpSession session) {
+    public String updateForm(@PathVariable Long index, Model model, HttpSession session) {
         User loginUser = (User) session.getAttribute("loginUser");
         Article article = articleService.findArticleIfOwner(loginUser, index);
 
@@ -68,7 +67,7 @@ public class ArticleController {
     }
 
     @PutMapping("/articles/{index}/update")
-    public String update(@PathVariable int index, ArticleForm articleForm, HttpSession session) {
+    public String update(@PathVariable Long index, ArticleForm articleForm, HttpSession session) {
         User loginUser = (User) session.getAttribute("loginUser");
         articleService.update(loginUser, index, articleForm);
